requirements.txt
django-heroku
Django==2.1.7
gunicorn 
psycopg2
django-paypal